Since 042stab083.2:
    I/O limits for containers (requires vzctl 4.6)
    iotop utility didn't work inside Containers (#2549)
    Added the ability to see in which Container the process was killed by OOM killer (PCLIN-32184)
    Fixed 'WARNING: at kernel/softirq.c:161 local_bh_enable_ip+0x7d/0xb0()' (PSBM-23346, #2526)
    A race in ploop_tracker_init() could lead to data corruption on Container's live migration (PSBM-22993)
    Kernel could crash on an attempt to access specific device (PSBM-21015)
    Multiple improvements and fixes regarding ploop images stability and performance (PSBM-23452, PSBM-23302, PSBM-22506)
    Other stability fixes (PSBM-23231, PSBM-20168, PSBM-23133, PSBM-22982, PSBM-22887, PSBM-22826, PSBM-22229, PSBM-22198, PSBM-21877, PSBM-21076, PSBM-20614, PSBM-18642)

For more details, please see changes for the relevant - testing kernels, starting from 042stab084.12 ( http://j.mp/UlmkGe )

http://j.mp/1oIDXsg

+++ kernel/kernel/cpt/cpt_dump.h
@@ -17,4 +17,4 @@ int cpt_check_unsupported(struct task_struct *tsk, struct cpt_context *ctx);
 
 extern unsigned long suspend_timeout_min;
 extern unsigned long suspend_timeout_max;
-extern unsigned long suspend_timeout;
+extern unsigned int suspend_timeout;
